NN: clean up only if installation finished
git-svn-id: http://svn.zentyal.org/zentyal/trunk@18763 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ extra/installer/data/second-boot.sh
@@ -1,13 +1,17 @@
 #!/bin/sh
 
-# Disable auto login once installation is done
-sed -i "s/auto_login.*/auto_login\tno/" /etc/slim.conf
+# If Zentyal is already installed...
+if ! [ -f '/var/lib/ebox/.first' ]
+then
+    # Disable auto login once installation is done
+    sed -i "s/auto_login.*/auto_login\tno/" /etc/slim.conf
 
-# Remove temporal local repository
-sed -i "/deb file.*ebox-packages/d" /etc/apt/sources.list
+    # Remove temporal local repository
+    sed -i "/deb file.*ebox-packages/d" /etc/apt/sources.list
 
-# Restore original rc.local and clean stuff
-mv /var/tmp/ebox/rc.local /etc/rc.local
-rm -rf /var/tmp/ebox
+    # Restore original rc.local and clean stuff
+    mv /var/tmp/ebox/rc.local /etc/rc.local
+    rm -rf /var/tmp/ebox
+fi
 
 exit 0
